Move AddPermissions component file
