+++ src/components/add-permissions.js
@@ -3,8 +3,8 @@ import { observer, inject } from 'mobx-react'
 import styled from 'styled-components'
 
 import { Field, TextInput, Button, RadioButton, DropDown } from '@aragon/ui'
-import { CheckButton } from '../../check-button'
-import { PermissionType } from '../../../stores/permissions-store'
+import { CheckButton } from './check-button'
+import { PermissionType } from '../stores/permissions-store'
