+++ stories/app.js
@@ -16,6 +16,10 @@ const configStore = new ConfigStore(datastore)
 const mainStore = new MainStore(datastore)
 const permissionsStore = new PermissionsStore(datastore, mainStore)
 
+datastore.createGroup("Group #1")
+datastore.createGroup("Lggkiwfj aef")
+datastore.createGroup("Group #32")
+
 
 aragonStoriesOf("Main App", module).add("Basic", () => (
   <Provider permissionsStore={permissionsStore} datastore={datastore} mainStore={mainStore} configStore={configStore}>
